(DBFLUTE-643){Java/C#}: DBFluteモジュール配下にアプリ情報を保持しない

+++ dbflute/src/main/java/org/seasar/dbflute/task/replaceschema/DfCreateSchemaTask.java
@@ -131,6 +131,10 @@ public class DfCreateSchemaTask extends DfAbstractReplaceSchemaTask {
         _log.info("* Initialize Schema (Additional Drop) *");
         _log.info("*                                     *");
         _log.info("* * * * * * * * * * * * * * * * * * * *");
+        if (_lazyConnection) {
+            _log.info("*Passed because it's a lazy connection");
+            return;
+        }
         for (Map<String, Object> additionalDropMap : additionalDropMapList) {
             final String dropSchema = getMyProperties().getAdditionalDropSchema(additionalDropMap);
             final String dropUrl = getMyProperties().getAdditionalDropUrl(additionalDropMap);
@@ -162,7 +166,6 @@ public class DfCreateSchemaTask extends DfAbstractReplaceSchemaTask {
         _log.info("* * * * * * * * * * *");
         if (_lazyConnection) {
             _log.info("*Passed because it's a lazy connection");
-            _log.info("");
             return;
         }
         final DfSchemaInitializer initializer = createSchemaInitializer(InitializeType.MAIN);
